+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `Add-ErAttribute` command to support attributes in ER diagrams.
+
 ## [0.6.0] - 2025-04-16
 
 ### Added